extract basic ad tracking plugin to separate file
